fix(ui): align with designs consolidation on master
#19316 moved MesheryPatterns/, DesignLifeCycle/, and configuratorComponents/
under ui/components/designs/. Update:
- designs/patterns/* and designs/lifecycle/* internal paths to ../../
  (sibling of components/) instead of ../ which the PR's commits used
  while assuming the legacy location.
- filters/FiltersGrid: import Grid.styles from designs/patterns/.
- layout/Navigator/NavigatorExtension, layout/NotificationCenter/{metadata,
  formatters/pattern_dryrun}: DesignLifeCycle/* -> designs/lifecycle/*.
- registry/ImportModelModal and registry/Stepper/FinishModelGenerateStep:
  @/components/DesignLifeCycle/* -> @/components/designs/lifecycle/*.
- eslint.config.js legacy allowlists: DesignLifeCycle/, MesheryPatterns/,
  configuratorComponents/ paths -> designs/{lifecycle,patterns,configurator}/.
