+++ ui/components/designs/lifecycle/SelectDeploymentTarget.tsx
@@ -15,7 +15,7 @@ import {
   useTheme,
 } from '@sistent/sistent';
 import { Loading, StepHeading } from './common';
-import { K8sContextConnectionChip } from '../layout/Header/Header';
+import { K8sContextConnectionChip } from '../../layout/Header/Header';
 import { createContext } from 'react';
 import { useContext } from 'react';
 import {
