+++ ui/components/designs/patterns/MesheryPatterns.tsx
@@ -45,13 +45,13 @@ import {
   Public as PublicIcon,
   Publish as PublishIcon,
 } from '@/assets/icons';
-import _PromptComponent from '../PromptComponent';
-import LoadingScreen from '../shared/LoadingState/LoadingComponent';
-import { FILE_OPS, MesheryPatternsCatalog, VISIBILITY } from '../../utils/Enum';
-import CloneIcon from '../../public/static/img/CloneIcon';
+import _PromptComponent from '../../PromptComponent';
+import LoadingScreen from '../../shared/LoadingState/LoadingComponent';
+import { FILE_OPS, MesheryPatternsCatalog, VISIBILITY } from '../../../utils/Enum';
+import CloneIcon from '../../../public/static/img/CloneIcon';
 import { useRouter } from 'next/router';
-import { RJSFModalWrapper } from '../shared/Modal/Modal';
-import downloadContent from '../../utils/fileDownloader';
+import { RJSFModalWrapper } from '../../shared/Modal/Modal';
+import downloadContent from '../../../utils/fileDownloader';
 import ConfigurationSubscription from '@/graphql/subscriptions/ConfigurationSubscription';
 import Pattern from '../../../public/static/img/drawer-icons/pattern_svg';
 import { useNotification } from '../../../utils/hooks/useNotification';
@@ -60,15 +60,15 @@ import _ from 'lodash';
 import { getMeshModels } from '../../../api/meshmodel';
 import { modifyRJSFSchema } from '../../../utils/utils';
 import { Edit as EditIcon } from '@/assets/icons';
-import { updateVisibleColumns } from '../../utils/responsive-column';
-import { useWindowDimensions } from '../../utils/dimension';
-import InfoModal from '../shared/Modal/Information/InfoModal';
+import { updateVisibleColumns } from '../../../utils/responsive-column';
+import { useWindowDimensions } from '../../../utils/dimension';
+import InfoModal from '../../shared/Modal/Information/InfoModal';
 import { InfoOutlined as InfoOutlinedIcon } from '@/assets/icons';
-import { DefaultTableCell, SortableTableCell } from '../connections/common';
-import DefaultError from '../General/error-404/index';
+import { DefaultTableCell, SortableTableCell } from '../../connections/common';
+import DefaultError from '../../General/error-404/index';
 import CAN from '@/utils/can';
 import { keys } from '@/utils/permission_constants';
-import ExportModal from '../shared/Modal/ExportModal';
+import ExportModal from '../../shared/Modal/ExportModal';
 import { useModal, Modal as SistentModal, ModalBody } from '@sistent/sistent';
 import PatternIcon from '@/assets/icons/Pattern';
 import DryRunIcon from '@/assets/icons/DryRunIcon';
