+++ ui/components/layout/Navigator/NavigatorExtension.tsx
@@ -16,9 +16,9 @@ import ExportModal from '../../shared/Modal/ExportModal';
 import { FormatStructuredData } from '../../DataFormatter';
 import { useFilterK8sContexts } from '@/utils/hooks/useKubernetesHook';
 import { useDynamicComponent } from '@/utils/context/dynamicContext';
-import { ValidateDesign } from '../../DesignLifeCycle/ValidateDesign';
-import { DryRunDesign } from '../../DesignLifeCycle/DryRun';
-import { DeployStepper, UnDeployStepper } from '../../DesignLifeCycle/DeployStepper';
+import { ValidateDesign } from '../../designs/lifecycle/ValidateDesign';
+import { DryRunDesign } from '../../designs/lifecycle/DryRun';
+import { DeployStepper, UnDeployStepper } from '../../designs/lifecycle/DeployStepper';
 import { designValidationMachine } from 'machines/validator/designValidator';
 import CAN from '@/utils/can';
 import { mesheryEventBus } from '@/utils/eventBus';
