+++ ui/components/layout/NotificationCenter/metadata.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { FormatStructuredData, SectionBody, reorderObjectProperties } from '../../DataFormatter';
 import { isEmptyAtAllDepths } from '../../../utils/objects';
 import { canTruncateDescription } from './notification';
-import { DeploymentSummaryFormatter } from '../../DesignLifeCycle/DeploymentSummary';
+import { DeploymentSummaryFormatter } from '../../designs/lifecycle/DeploymentSummary';
 import { EVENT_TYPE, eventDetailFormatterKey } from './constants';
 import { TitleLink, DataToFileLink, EmptyState } from './formatters/common';
 import { ErrorMetadataFormatter } from './formatters/error';
