+++ ui/eslint.config.js
@@ -36,10 +36,10 @@ const legacyRestrictedImportOffenders = [
   'components/Dashboard/resources/resources-sub-menu.tsx',
   'components/Dashboard/widgets/HoneyComb/HoneyCombComponent.tsx',
   'components/DataFormatter/index.tsx',
-  'components/DesignLifeCycle/DeploymentSummary.tsx',
-  'components/DesignLifeCycle/DryRun.tsx',
-  'components/DesignLifeCycle/ValidateDesign.tsx',
-  'components/DesignLifeCycle/styles.tsx',
+  'components/designs/lifecycle/DeploymentSummary.tsx',
+  'components/designs/lifecycle/DryRun.tsx',
+  'components/designs/lifecycle/ValidateDesign.tsx',
+  'components/designs/lifecycle/styles.tsx',
   'components/layout/Header/Header.tsx',
   'components/environments/environment-card.tsx',
   'components/environments/index.tsx',
@@ -102,10 +102,10 @@ const legacyLiteralColorOffenders = [
   'components/MesheryMeshInterface/PatternService/RJSFCustomComponents/CustomBaseInput.tsx',
   'components/MesheryMeshInterface/PatternService/RJSFCustomComponents/ObjectFieldTemplate.tsx',
   'components/MesheryMeshInterface/PatternServiceForm.tsx',
-  'components/MesheryPatterns/MesheryPatternCard.tsx',
-  'components/MesheryPatterns/MesheryPatternGridView.tsx',
-  'components/MesheryPatterns/MesheryPatterns.tsx',
-  'components/MesheryPatterns/style.tsx',
+  'components/designs/patterns/MesheryPatternCard.tsx',
+  'components/designs/patterns/MesheryPatternGridView.tsx',
+  'components/designs/patterns/MesheryPatterns.tsx',
+  'components/designs/patterns/style.tsx',
   'components/layout/Navigator/NavigatorExtension.tsx',
   'components/layout/NotificationCenter/constants.tsx',
   'components/layout/NotificationCenter/formatters/relationship_evaluation.tsx',
